requirements.txt
flask==3.0.3
flask-restx==1.3.0
python-dotenv==1.0.1
#flask-restful==0.3.10
flask-sqlalchemy==3.1.1
flask-jwt-extended==4.6.0
flask-mail==0.9.1
httpx==0.24.1

# Testing dependencies
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0
